fix(combinator): Pull in sequence
