+++ examples/arithmetic/parser.rs
@@ -4,8 +4,8 @@ use winnow::prelude::*;
 use winnow::{
     branch::alt,
     character::{digit1 as digits, space0 as spaces},
+    combinator::delimited,
     multi::fold_many0,
-    sequence::delimited,
     token::one_of,
     IResult,
 };
