+++ examples/ini/parser.rs
@@ -5,8 +5,8 @@ use winnow::prelude::*;
 use winnow::{
     character::{alphanumeric1 as alphanumeric, multispace0 as multispace, space0 as space},
     combinator::opt,
+    combinator::{delimited, separated_pair, terminated},
     multi::many0,
-    sequence::{delimited, separated_pair, terminated},
     token::take_while0,
 };
