+++ examples/ini/parser_str.rs
@@ -4,8 +4,8 @@ use winnow::prelude::*;
 use winnow::{
     character::{alphanumeric1 as alphanumeric, space0 as space},
     combinator::opt,
+    combinator::{delimited, terminated},
     multi::many0,
-    sequence::{delimited, terminated},
     token::{take_till0, take_while0, take_while1},
 };
