+++ examples/json/parser.rs
@@ -6,9 +6,9 @@ use winnow::{
     branch::alt,
     character::float,
     combinator::cut_err,
+    combinator::{delimited, preceded, separated_pair, terminated},
     error::{ContextError, ParseError},
     multi::{fold_many0, separated0},
-    sequence::{delimited, preceded, separated_pair, terminated},
     token::{any, none_of, take, take_while0},
 };
