+++ examples/json/parser_partial.rs
@@ -6,9 +6,9 @@ use winnow::{
     branch::alt,
     character::float,
     combinator::{cut_err, rest},
+    combinator::{delimited, preceded, separated_pair, terminated},
     error::{ContextError, ParseError},
     multi::{fold_many0, separated0},
-    sequence::{delimited, preceded, separated_pair, terminated},
     stream::Partial,
     token::{any, none_of, take, take_while0},
 };
