+++ examples/json_iterator.rs
@@ -5,9 +5,9 @@ use winnow::{
     branch::alt,
     character::{alphanumeric1 as alphanumeric, escaped, float},
     combinator::cut_err,
+    combinator::{preceded, separated_pair, terminated},
     error::ParseError,
     multi::separated0,
-    sequence::{preceded, separated_pair, terminated},
     token::one_of,
     token::{tag, take_while0},
     IResult,
