+++ tests/testsuite/reborrow_fold.rs
@@ -3,9 +3,9 @@
 
 use std::str;
 
+use winnow::combinator::delimited;
 use winnow::multi::fold_many0;
 use winnow::prelude::*;
-use winnow::sequence::delimited;
 use winnow::token::take_till1;
 use winnow::IResult;
